+++ source/databricks/package/calculator_job.py
@@ -22,7 +22,7 @@ from pyspark.sql import DataFrame, SparkSession
 from pyspark.sql.functions import col
 from pyspark.sql.types import Row
 from configargparse import argparse
-from package.constants import Colname
+from package.calculation_output_writer import CalculationOutputWriter
 from package import (
     calculate_balance_fixing,
     db_logging,
@@ -62,17 +62,6 @@ def _get_valid_args_or_throw(command_line_args: list[str]) -> argparse.Namespace
     return args
 
 
-def write_basis_data_to_csv(data_df: DataFrame, path: str) -> None:
-    (
-        data_df.withColumnRenamed("GridAreaCode", "grid_area")
-        .repartition("grid_area")
-        .write.mode("overwrite")
-        .partitionBy("grid_area", "gln")
-        .option("header", True)
-        .csv(path)
-    )
-
-
 def _start_calculator(spark: SparkSession, args: CalculatorArgs) -> None:
     timeseries_points_df = (
         spark.read.option("mode", "FAILFAST")
@@ -100,66 +89,17 @@ def _start_calculator(spark: SparkSession, args: CalculatorArgs) -> None:
         args.batch_period_end_datetime,
     )
 
-    (
-        non_profiled_consumption_per_ga_and_es,
-        production_per_ga_df,
-        timeseries_basis_data_df,
-        master_basis_data_df,
-    ) = calculate_balance_fixing(
+    calculation_output_writer = CalculationOutputWriter(
+        args.batch_id, args.process_results_path
+    )
+
+    calculate_balance_fixing(
+        calculation_output_writer,
         metering_point_periods_df,
         timeseries_points_df,
         args.batch_period_start_datetime,
         args.batch_period_end_datetime,
         args.time_zone,
-    )
-
-    (timeseries_quarter_df, timeseries_hour_df) = timeseries_basis_data_df
-    debug("timeseries basis data df_hour", timeseries_hour_df)
-    debug("timeseries basis data df_quarter", timeseries_quarter_df)
-    debug("master basis data", master_basis_data_df)
-
-    write_basis_data_to_csv(
-        timeseries_quarter_df,
-        f"{args.process_results_path}/batch_id={args.batch_id}/basis_data/time_series_quarter",
-    )
-
-    write_basis_data_to_csv(
-        timeseries_hour_df,
-        f"{args.process_results_path}/batch_id={args.batch_id}/basis_data/time_series_hour",
-    )
-
-    write_basis_data_to_csv(
-        master_basis_data_df,
-        f"{args.process_results_path}/batch_id={args.batch_id}/basis_data/master_basis_data",
-    )
-
-    path = f"{args.process_results_path}/batch_id={args.batch_id}/result"
-
-    # First repartition to co-locate all rows for a grid area on a single executor.
-    # This ensures that only one file is being written/created for each grid area
-    # When writing/creating the files. The partition by creates a folder for each grid area.
-
-    # Total production
-    (
-        production_per_ga_df.repartition("grid_area")
-        .write.mode("overwrite")
-        .partitionBy("grid_area", Colname.gln, Colname.time_series_type)
-        .json(path)
-    )
-
-    # Non-profiled consumption
-    (
-        non_profiled_consumption_per_ga_and_es.repartition("grid_area")
-        .write.mode("append")
-        .partitionBy("grid_area", Colname.gln, Colname.time_series_type)
-        .json(path)
-    )
-
-    (
-        non_profiled_consumption_per_ga_and_es.repartition("grid_area")
-        .write.mode("append")
-        .partitionBy("grid_area", Colname.gln, Colname.time_series_type)
-        .json(path, compression="gzip")
     )