+++ source/dotnet/Contracts/ProcessCompleted.cs
@@ -12,12 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Infrastructure.Integration;
+// DO NOT CHANGE - this namespace is exposed publicly in the Contracts package
+namespace Energinet.DataHub.Wholesale.Contracts.Events;
 
 public partial class ProcessCompleted
 {
     /// <summary>
-    /// The message type suitable for ServiceBus subscription filtering in accordance with ADR-008.
+    /// The message type for balance-fixing-completed transport message meta data in accordance with ADR-008.
     /// </summary>
-    public const string MessageType = "ProcessCompleted";
+    public const string BalanceFixingProcessType = "BalanceFixingCompleted";
 }