+++ source/dotnet/Services/Infrastructure/Batches/ServiceCollectionExtensions.cs
@@ -13,7 +13,9 @@
 // limitations under the License.
 
 using Azure.Messaging.ServiceBus;
+using Energinet.DataHub.Core.JsonSerialization;
 using Energinet.DataHub.Wholesale.Application.Batches;
+using Energinet.DataHub.Wholesale.Application.Infrastructure;
 using Energinet.DataHub.Wholesale.Infrastructure.Registration;
 using Energinet.DataHub.Wholesale.Infrastructure.ServiceBus;
 using Microsoft.Extensions.DependencyInjection;
@@ -25,7 +27,8 @@ public static class ServiceCollectionExtensions
     public static IServiceCollection AddBatchCompletedPublisher(
         this IServiceCollection serviceCollection,
         string serviceBusConnectionString,
-        string batchCompletedTopicName)
+        string batchCompletedTopicName,
+        string messageType)
     {
         serviceCollection.AddScoped<IBatchCompletedPublisher>(provider =>
         {
@@ -33,7 +36,8 @@ public static class ServiceCollectionExtensions
                 .GetRequiredService<TargetedSingleton<ServiceBusSender, BatchCompletedPublisher>>()
                 .Instance;
             var factory = provider.GetRequiredService<IServiceBusMessageFactory>();
-            return new BatchCompletedPublisher(sender, factory);
+            var serializer = provider.GetRequiredService<IJsonSerializer>();
+            return new BatchCompletedPublisher(sender, factory, messageType, serializer);
         });
 
         if (serviceCollection.All(x => x.ServiceType != typeof(ServiceBusClient)))