+++ source/dotnet/Services/Infrastructure/Integration/ProcessCompletedIntegrationEventMapper.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Application.Processes;
-using Energinet.DataHub.Wholesale.Contracts;
+using Energinet.DataHub.Wholesale.Contracts.Events;
 
 namespace Energinet.DataHub.Wholesale.Infrastructure.Integration;
 
@@ -24,18 +24,9 @@ public class ProcessCompletedIntegrationEventMapper : IProcessCompletedIntegrati
         return new ProcessCompleted
         {
             BatchId = processCompletedEvent.BatchId.ToString(),
-            ProcessType = GetProcessType(processCompletedEvent),
             GridAreaCode = processCompletedEvent.GridAreaCode,
             PeriodStartUtc = processCompletedEvent.PeriodStart.ToTimestamp(),
             PeriodEndUtc = processCompletedEvent.PeriodEnd.ToTimestamp(),
         };
     }
-
-    private static ProcessCompleted.Types.ProcessType GetProcessType(ProcessCompletedEventDto processCompletedEvent) =>
-        processCompletedEvent.ProcessType switch
-        {
-            ProcessType.BalanceFixing => ProcessCompleted.Types.ProcessType.PtBalancefixing,
-            _ => throw new NotImplementedException(
-                $"Cannot map process type '{processCompletedEvent.ProcessType.ToString()}'"),
-        };
 }