+++ source/dotnet/Packages/Client.Documentation/release-notes/release-notes.md
@@ -1,5 +1,13 @@
 # Wholesale Client Package Release notes
 
+## Version 4.2.0
+
+Add quality to process time series type results. See `TimeSeriesPointDto.Quality`.
+
+## Version 4.1.1
+
+Internal refactoring. No changed behavior.
+
 ## Version 4.1.0
 
 Changes to `BatchDtoV2`.